refactoring registry app, and add new app creation
