+++ services/user/Registry/service/src/constants.rs
@@ -1,0 +1,6 @@
+// An App can't have more than 3 tags
+pub const MAX_APP_TAGS: usize = 3;
+pub const MAX_TAG_LENGTH: usize = 30;
+pub const MAX_APP_NAME_LENGTH: usize = 30;
+pub const MAX_APP_SHORT_DESCRIPTION_LENGTH: usize = 100;
+pub const MAX_APP_LONG_DESCRIPTION_LENGTH: usize = 1000;
